cmake: fix header configuration detection
file(COPY is woefully insufficient for installing headers -
it doesn't track changes, and doesn't rerun at reconfigure.
Essentially this means that if you edit a header, make won't
copy the file, and you will get confusing errors pointing to the
old version of the header that shouldn't exist anymore.

file(GLOB_RECURSE has a helpful option CONFIGURE_DEPENDS that will
properly detect changes. In conjunction with configure_file, this
means we can now rerun make in a development workflow with fewer headaches.

Please note that cmake globbing is nonetheless a bit dodgy,
and the only way to have a truly robust workflow in cmake
is to explicitly list all headers + sources :/