journal: added splitter class
A Splitter is a plumbing Device which maps one input to many outputs.
It appears the same as a typical Device and exposes the same interface,
but actually contains a collection of output Devices.
Any operations done on a Splitter are mapped to all outputs.

This was inspired by spdlog's loggers and sinks.
(spdlog::logger ~ journal::channel, spdlog::sink ~ journal::device_t)
A spdlog::logger contains a set of sinks,
whereas a journal::channel has exactly one Device.
But journal's architecture is sufficiently general that
we can emulate this behavior!

Since each Device must be given a name, but Splitters probably shouldn't
be referred to by name since they are simply plumbing, I set them up to
autogenerate a unique internal ID. I considered creating e.g. an
AnonymousDevice base class for Device, but deferred since this would be
intrusive. Does it make sense for certain types of "plumbing" Devices
to exist beyond the purview of Chronicler and friends?

I'm planning to create another Device based on Splitter which splits to
the console and one or more files, which will be named Tee.
Creating a Tee was my original motivation for this class, but I thought
it appropriate to first create the generic plumbing class first.
I imagine Tee would capture the majority of my usage of Splitter, but
it's quite possible that there are more uses for such a general class.