lib/pyre/algebra: diagonal matrices are now rank 2 tensors with diagonal packing
Operators defined on Tensors (e.g. operator==, operator+, operator-) are now consistently implemented in a packing-independent fashion, i.e. using indices as opposed to offsets. These operators are now implemented with iterators on the {layout}, which is private. Iterators on {Tensors} wrapping the iterators of the layout are yet to be implemented.

NOTE: The diagonal packing has the peculiarity that the number of total entries in the {Tensor}, i.e. N, does not match the number of cells in the {layout}. In fact, the {layout} of a diagonal packing needs an extra entry to store the off-diagonal zero.