lib/pyre/tensor: moved literal tensors out of class tensors
This edit fixes a {clang} complaint. According to {clang}, these tensors cannot be a {constexpr} member of class {Tensor} because their type is {Tensor} and the type {Tensor} is not a complete type within the {Tensor} class scope, therefore they cannot be literals (according to the standard, incomplete types cannot be literal types).